requirements.txt
transformers
crewai
python-dotenv
linkedin-api
fpdf
groq
crewai_tools
